% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{al_particulate}
\alias{al_particulate}
\title{Calculate aluminum particulate}
\usage{
al_particulate(labdat)
}
\arguments{
\item{labdat}{dataframe containing the parsed lab data}
}
\value{
numeric vector. The calculated values
}
\description{
Calculate aluminum particulate, equal to total aluminum - dissolved aluminum
}
