% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{alum_DOC_ratio}
\alias{alum_DOC_ratio}
\title{Calculate alum to DOC ratio}
\usage{
alum_DOC_ratio(column_names, labdat)
}
\arguments{
\item{column_names}{string vector defined in apply_calculations. The desired final
columns for each calculated dataframe}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
numeric vector. The calculated values
}
\description{
Calculate alum to DOC ratio, equal to Alum/DOC
}
