% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{alumDOC_ratio}
\alias{alumDOC_ratio}
\title{Calculate Alum to DOC ratio}
\usage{
alumDOC_ratio(labdat)
}
\arguments{
\item{labdat}{routine lab dat file}
}
\value{
calculated alum:DOC ratio
}
\description{
Calculate Alum to DOC ratio
}
