% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{alumDOC_stoich}
\alias{alumDOC_stoich}
\title{Calculate Alum to DOC stoichiometry}
\usage{
alumDOC_stoich(labdat)
}
\arguments{
\item{labdat}{routine lab dat file}
}
\value{
calculated alum:DOC stoichiometry
}
\description{
Calculate Alum to DOC stoichiometry
}
