% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_newdata}
\alias{append_newdata}
\title{Add new data to dataset}
\usage{
append_newdata(labdat, new_data)
}
\arguments{
\item{labdat}{labdat datafile}

\item{new_data}{new data file}
}
\value{

}
\description{
Add new data to dataset
}
