% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{build_database}
\alias{build_database}
\title{Build database from scraped BPWTP routine lab data}
\usage{
build_database(
  datadir = "data/labdat_datafiles",
  outdir = "output",
  outfile = "test_output.csv",
  save_output = FALSE
)
}
\arguments{
\item{datadir}{data directory}

\item{outdir}{output directory}

\item{outfile}{out file name}

\item{save_output}{toggled option to save data file}
}
\value{
single data file with all historical data
}
\description{
Build database from scraped BPWTP routine lab data
}
\examples{
# build_database(datadir = "data/labdat_files", outdir = "data", outfile = "testfile.csv")
}
