% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_TDS}
\alias{calc_TDS}
\title{Total Dissolved Solids Calculation (Raw Water & Clearwell)}
\usage{
calc_TDS(labdat)
}
\arguments{
\item{labdat}{Weekly routine sampling data.}
}
\value{
Dataframe containing datasheet (RawWater, ClearWell), station (Raw,
Clearwell), datetimes, parameter (calc.TDS_mg.L), and result (raw
water and clearwell TDS concentration in mg/L).
}
\description{
From 2014—2017, and in addition to a direct TDS measurement, TDS was
calculated as the sum of bicarbonate, carbonate, calcium, magnesium, sodium,
potassium, sulphate, chloride, and silica ion concentrations (mg/L).
}
\details{
TDS (mg/L) = Bicarbonate + Carbonate + Calcium + Magnesium + Sodium +
Potassium + + Sulphate + Chloride + Silica.
}
\examples{
calc_TDS(labdat = labdat)

}
