% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_value}
\alias{calculate_value}
\title{Calculate value}
\usage{
calculate_value(
  column_names,
  labdat,
  parms,
  station_list = NULL,
  fn,
  tbl_parameter,
  tbl_unit,
  tbl_parm_unit,
  tbl_parm_tag,
  round_to
)
}
\arguments{
\item{column_names}{constant string vector. The desired column names in order}

\item{labdat}{dataframe containing the parsed lab data}

\item{parms}{string vector. The names of the columns required for the
calculation of interest}

\item{station_list}{string vector. The stations whose parameters are required
for the calculation of interest. Relevant particularly for calculations that
use parameters that exist for multiple stations but are only required for
certain stations. Default: NULL}

\item{fn}{function to apply to \code{parms}}

\item{tbl_parameter}{string. New calculated parameter name}

\item{tbl_unit}{string.  New calculated parameter unit}

\item{tbl_parm_unit}{string. New calculated parameter parm_unit}

\item{tbl_parm_tag}{string. New calculated parameter tag}

\item{round_to}{numeric. Number of desired decimal places}
}
\value{
dataframe containing all required columns, including the calculated
result.
Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate a specific value given the specified function
}
