% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_missing_params}
\alias{check_missing_params}
\title{Identify which parameters are missing in parameters.xlsx}
\usage{
check_missing_params(path_to_labdat_file, path_to_parameters)
}
\arguments{
\item{path_to_labdat_file}{string. Path to the lab data file to be inserted
into the database file}

\item{path_to_parameters}{string. Path to the .xlsx file that contains the
desired parameters and associated info (station, units, corrected units...)}
}
\value{
Print message showing user the next action steps
}
\description{
Identify the user that certain rows from the lab data weekly data sheet
are not read in by the script.
}
