% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_parameters_rwcw}
\alias{check_parameters_rwcw}
\title{Identify which Raw/Clearwell parameters are missing in parameters.xlsx}
\usage{
check_parameters_rwcw(path_to_labdat_file, path_to_parameters)
}
\arguments{
\item{path_to_labdat_file}{string. Full path (from root directory) to the lab
data file to be inserted into the database file}

\item{path_to_parameters}{string. Full path (from root directory) to the
.xlsx file that contains the desired parameters and associated info
(station, units, corrected units...)}
}
\value{
Print message showing user the next action steps
}
\description{
Inform the user that certain rows from the lab data weekly data sheet
are not read in by the script.
}
