% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_parameters_setup}
\alias{check_parameters_setup}
\title{Check parameters.xlsx}
\usage{
check_parameters_setup(labdat_parameters)
}
\arguments{
\item{labdat_parameters}{dataframe. Output from check_parameters_content}
}
\value{
dataframe containing the corrected parameter information
}
\description{
Check parameters.xlsx column names and ensure that duplicate rows are removed
}
