% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_scraped_data}
\alias{check_scraped_data}
\title{Check scraped data}
\usage{
check_scraped_data(new_data, labdat_parameters)
}
\arguments{
\item{new_data}{dataframe. Output from scrape_labdatxls}

\item{labdat_parameters}{dataframe. Slightly processed parameters.xlsx data}
}
\value{
Print message showing user the next action steps
}
\description{
Check data read in with scrape_labdatxls by and scrape_docprofiles. Ensure
that scrapes were done correctly such that there is only one set of data
values for each combination of datasheet, station, parameter, unit
}
