% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{convert_bdl}
\alias{convert_bdl}
\title{Replace below detection limit values with}
\usage{
convert_bdl(labdat, replacement = "0")
}
\arguments{
\item{replacement}{}
}
\value{

}
\description{
Replace below detection limit values with
}
