% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{convert_biocounts}
\alias{convert_biocounts}
\title{Convert biological data}
\usage{
convert_biocounts(labdat)
}
\arguments{
\item{labdat}{labdat file}
}
\value{
converted values
}
\description{
Convert biological data
}
