% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{convert_UV254}
\alias{convert_UV254}
\title{Convert UV 254 values.}
\usage{
convert_UV254(labdat)
}
\arguments{
\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing the parsed lab data with the required UV 254
data values and units converted to Abs 1cm as required
}
\description{
Convert UV 254 values from Abs 10cm to Abs 1cm by dividing values by 10
}
