% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{create_doc_colnames}
\alias{create_doc_colnames}
\title{Create DOC colnames}
\usage{
create_doc_colnames(val1, val2, val3)
}
\arguments{
\item{val1}{string. Value in 1st header row}

\item{val2}{string. Value in 2nd header row}

\item{val3}{string. Value in 3rd header row, if there are three header rows.
NA otherwise}
}
\value{
string. Column name
}
\description{
Use all available header row info to create full column names for DOC profile
data
}
