% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{decimal_places}
\alias{decimal_places}
\title{Determine number of decimal places}
\usage{
decimal_places(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
numeric value. The number of decimal places of the value. Does not
consider 1.00 to have 2 decimal places but 1
}
\description{
Determine the number of decimal places of a numeric value
}
