% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{det_cols}
\alias{det_cols}
\title{Calculate vector of desired columns}
\usage{
det_cols(parms)
}
\arguments{
\item{parms}{string vector. The names of the columns specific to the
calculation of interest}
}
\value{
string vector. The names of all the columns of interest for a
calculation of interest (including datasheet, sheet year, station, date, and
whatever other columns are specific to the calculation of interest)
}
\description{
Create a numeric vector to store the names of the columns required for the
calculation of interest. Used to later reorganize data and to add columns
that are missing from a dataset but that are required for a calculation.
}
