% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{determine_NA_NC}
\alias{determine_NA_NC}
\title{Determine which calculated values will are valid}
\usage{
determine_NA_NC(df, col1, col2, include_between = F, col_name)
}
\arguments{
\item{df}{dataframe containing measured values required for a specific calculation}

\item{col1}{first parameter required for the calculation}

\item{col2}{last parameter required for the calculation}

\item{include_between}{boolean value, defaults to TRUE
Do there exist columns between col1 and col2 that must be examined for the
existence of valid numbers?
If FALSE, no columns other than col1 and col2 must be examined.}

\item{col_name}{string used to assign a name to the column that stores the
result of \code{determine_NA_NC()}. If not specified, column is given the name
\code{result}}
}
\value{
A dataframe identical to the input dataframe with one additional
column containing the result of \code{determine_NA_NC()} and named using
col_name. Values are assigned as follows:
\itemize{
\item When all of the parameters required for the calculation are missing
values, assign NA
\item When one or more (but not all) of the parameters required for the
calculation are missing values, assign NaN
\item When all of the parameters required for the calculation are present
values, assign 999 as a placeholder to indicate that a value should be
calculated
}
}
\description{
Classifies values as NA, NaN, or 999 depending on whether or not the measured
values required for that calculation exist for none, some, or all of the
parameters.
}
