% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{dissolved_solids}
\alias{dissolved_solids}
\title{Calculate the dissolved solids
Calculate the dissolved solids, equal to the sum of bicarbonate, carbonate,
calcium, magnesium, sodium, potassium, sulphate, chloride, silica}
\usage{
dissolved_solids(column_names, labdat)
}
\arguments{
\item{column_names}{constant string vector. The desired column names in order}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate the dissolved solids
Calculate the dissolved solids, equal to the sum of bicarbonate, carbonate,
calcium, magnesium, sodium, potassium, sulphate, chloride, silica
}
