% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{DOCremoval_coag}
\alias{DOCremoval_coag}
\title{Calculate DOC removal by coagulation}
\usage{
DOCremoval_coag(labdat)
}
\arguments{
\item{labdat}{routine labdat dataset}
}
\value{
calculated odour removal by coagulation
}
\description{
Calculate DOC removal by coagulation
}
\examples{
 #DOCremoval_GACfilt(labdat)
}
