% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{DOCremoval_filt}
\alias{DOCremoval_filt}
\title{Calculate DOC removal by filtration}
\usage{
DOCremoval_filt(labdat)
}
\arguments{
\item{labdat}{routine labdat dataset}
}
\value{
calculated odour removal by coagulation
}
\description{
Calculate DOC removal by filtration
}
\examples{
 #DOCremoval_GACfilt(labdat)
}
