% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{DOCremoval_GACfilt}
\alias{DOCremoval_GACfilt}
\title{Calculate DOC removal my GAC}
\usage{
DOCremoval_GACfilt(labdat)
}
\arguments{
\item{labdat}{routine labdat dataset}
}
\value{
calculated odour removal by coagulation
}
\description{
Calculate DOC removal my GAC
}
\examples{
 #DOCremoval_GACfilt(labdat)
}
