% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_al}
\alias{filter_al}
\title{Select Al rows}
\usage{
filter_al(labdat_parameters)
}
\arguments{
\item{labdat_parameters}{dataframe. Output from check_parameters_content}
}
\value{
dataframe. All Al-related parameters
}
\description{
Select Al rows
}
