% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_ions}
\alias{filter_ions}
\title{Select ion rows}
\usage{
filter_ions(labdat_parameters)
}
\arguments{
\item{labdat_parameters}{dataframe. Output from check_parameters_content}
}
\value{
dataframe. All ion-related parameters
}
\description{
Select ion rows
}
