% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_thms}
\alias{filter_thms}
\title{Select THM rows}
\usage{
filter_thms(labdat_parameters)
}
\arguments{
\item{labdat_parameters}{dataframe. Output from check_parameters_content}
}
\value{
dataframe. All THM-related parameters
}
\description{
Select THM rows
}
