% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{handle_missing_cols}
\alias{handle_missing_cols}
\title{Add columns}
\usage{
handle_missing_cols(df, cols)
}
\arguments{
\item{df}{dataframe}

\item{cols}{A vector containing the desired columns in the desired order,
setup as c(colname = NA_real_, colname = NA_real_, ...)}
}
\value{
Dataframe with desired columns in the desired order
}
\description{
Add columns that are required for function processing in the case that
required columns are missing from the input dataframe.
}
