% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_calculations.R
\name{historical_extremes}
\alias{historical_extremes}
\title{Calculate historical extremes}
\usage{
historical_extremes(
  df = "",
  sampling_station = "",
  grouping = "",
  by_month = T
)
}
\arguments{
\item{by_month}{}
}
\value{

}
\description{
Calculate historical extremes
}
