% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{ion_balance}
\alias{ion_balance}
\title{Calculate ion balance}
\usage{
ion_balance(labdat)
}
\arguments{
\item{labdat}{routine labdat}
}
\value{
calculated ion balance
}
\description{
Calculate ion balance
}
