% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{langelier_SatIndex}
\alias{langelier_SatIndex}
\title{Calculate langelier saturation indeces}
\usage{
langelier_SatIndex(labdat)
}
\arguments{
\item{labdat}{routine lab data}
}
\value{
calculated lsi values
}
\description{
Calculate langelier saturation indeces
}
