% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{odourremoval_coag}
\alias{odourremoval_coag}
\title{Calculate odour removal during coagulation}
\usage{
odourremoval_coag(labdat)
}
\arguments{
\item{labdat}{routine labdat dataset}
}
\value{
calculated odour removal by coagulation
}
\description{
Calculate odour removal during coagulation
}
\examples{
 #odourremoval_coag(labdat)
}
