% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{odourremoval_filt}
\alias{odourremoval_filt}
\title{Calculate odour removal during filtration}
\usage{
odourremoval_filt(labdat)
}
\arguments{
\item{labdat}{routine labdat dataset}
}
\value{
calculated odour removal by filtration
}
\description{
Calculate odour removal during filtration
}
\examples{
 #odourremoval_filt(labdat)
}
