% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{PACl_DAE}
\alias{PACl_DAE}
\title{Calculate PACl DAE}
\usage{
PACl_DAE(column_names, PACl_values)
}
\arguments{
\item{column_names}{constant string vector. The desired column names in order}

\item{PACl_values}{numeric vector. The values calculated by \code{PACl()}}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate PACl as dry alum equivalent, equal to PACl\*2.07
}
