% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{parm_fig}
\alias{parm_fig}
\title{Plot parameter}
\usage{
parm_fig(df = "", year_select = "", var = "", outdir)
}
\arguments{
\item{df}{dataset}

\item{year_select}{year of interest}

\item{var}{parameter variable}

\item{outdir}{location to save output}
}
\value{
ggplot
}
\description{
Plot parameter figure with trend and boxplot distribution
}
