% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{percent_yield}
\alias{percent_yield}
\title{Calculate percent yield}
\usage{
percent_yield(pre, post)
}
\arguments{
\item{pre}{numeric value. First measurement}

\item{post}{numeric value. Last measurement}
}
\value{
numeric value. Calculated percent yield
}
\description{
Calculates the percent yield, equal to (pre-post)/pre*100
}
