% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_labdat.R
\name{prepare_labdat}
\alias{prepare_labdat}
\title{Prepare lab data}
\usage{
prepare_labdat(path_to_labdat_file, path_to_db_file, path_to_parameters)
}
\arguments{
\item{path_to_labdat_file}{string. Full path (from root directory) to the lab
data file to be inserted into the database file}

\item{path_to_db_file}{string. Full path (from root directory) to the .csv
database file if it exists, and desired path to a .csv database file if it
does not exist}

\item{path_to_parameters}{string. Full path (from root directory) to the
.xlsx file that contains the desired parameters and associated info
(station, units, corrected units...)}
}
\value{
dataframe containing the processed lab data that has been inserted
into the database file
}
\description{
Prepare lab data for incorporation into database. Extract the data that is
not yet in the database, format it, and insert into the database file.
}
