% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_db}
\alias{read_db}
\title{Read the csv containing the database}
\usage{
read_db(path_to_db_file)
}
\arguments{
\item{path_to_db_file}{string. Full path (from root directory) to the .csv
database file if it exists, and desired path to a .csv database file if it
does not exist}
}
\value{
dataframe containing the database, if it exists, or NULL if it does not
}
\description{
Read the csv containing all of the processed labdat data, if it exists. If it
does not exist, it will later be created
}
