% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_doc}
\alias{read_doc}
\title{Read WTP DOC Data sheet of lab data workbook}
\usage{
read_doc(path_to_labdat_file)
}
\arguments{
\item{path_to_labdat_file}{string. Full path (from root directory) to the lab
data file to be inserted into the database file}
}
\value{
dataframe containing the DOC data with edited column headers
}
\description{
Read WTP DOC Data sheet of lab data workbook
}
