% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_labdat}
\alias{read_labdat}
\title{Read routine lab data file}
\usage{
read_labdat(datadir = "data", labdat_file = "")
}
\arguments{
\item{datadir}{data directory}

\item{labdat_file}{lab data filename}
}
\value{
data table with replaced bdl values
}
\description{
Read routine lab data file
}
\examples{
# read_labdat(datadir = "data", labdat_file = "BPWTP_labdat.csv")
}
