% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_parameters}
\alias{read_parameters}
\title{Read parameters.xlsx file}
\usage{
read_parameters(path_to_parameters)
}
\arguments{
\item{path_to_parameters}{string. Full path (from root directory) to the
.xlsx file that contains the desired parameters and associated info
(station, units, corrected units...)}
}
\value{
dataframe containing the parameters info
}
\description{
Read parameters.xlsx file
}
