% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_generate.R
\name{report_generate}
\alias{report_generate}
\title{Generate bpwtp report}
\usage{
report_generate(path_to_db_file, outdir, selected_year, selected_month)
}
\arguments{
\item{path_to_db_file}{string. Path to the database csv file}

\item{outdir}{string. Path to the folder in which to deposit the outputs}

\item{selected_year}{numeric. Year of interest to compare to historical data}

\item{selected_month}{string. Abbreviated 3 letter month label. Ex "Jan"}
}
\value{
.Rmd
}
\description{
Generate bpwtp report
}
