% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{round_values}
\alias{round_values}
\title{Round values stored in result_org and result}
\usage{
round_values(labdat)
}
\arguments{
\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing the parsed lab data where result_org and result
are recorded with the correct number of decimal points
}
\description{
Round values stored in result_org and result as indicated in the process
document. Required as there exists an issue with number storage such that
when values are read in, some have excess decimal places. Especially
pertinent for values calculated within the lab data sheet
}
