% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_clearwell_al}
\alias{scrape_clearwell_al}
\title{Scrape Clearwell aluminum data}
\usage{
scrape_clearwell_al(spreadsheet, clearwell_start, labdat_parameters)
}
\arguments{
\item{spreadsheet}{dataframe. All data in the Weekly sheet of the lab data file}

\item{clearwell_start}{numeric value. The start of the Clearwell data as
indicated by the phrase "CLEAR WELL" in the Parameters column of the Weekly
data}

\item{labdat_parameters}{dataframe. Slightly processed parameters.xlsx data}
}
\value{
data frame of the Clearwell aluminum data
}
\description{
Read in Clearwell aluminum data. Identify parameters to read in based on
parameters.xlsx
}
