% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_docprofiles}
\alias{scrape_docprofiles}
\title{Scrape DOC profiles}
\usage{
scrape_docprofiles(labdat_filename)
}
\arguments{
\item{labdat_filename}{provided file name for the active data file}
}
\value{

}
\description{
Scrape DOC profiles
}
\examples{
#scrape_docprofiles(labdat_filename)
}
