% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_labdatxls}
\alias{scrape_labdatxls}
\title{Scrape data from BPWTP routine lab data}
\usage{
scrape_labdatxls(weekly_data, labdat_parameters)
}
\arguments{
\item{weekly_data}{dataframe. Raw weekly data outputted from read_weekly()}

\item{labdat_parameters}{dataframe. Output from check_parameters_content}
}
\value{
data frame of all the data contained in the Weekly lab data file
}
\description{
Read in all data contained in the Weekly (1st) sheet of the lab data file.
}
