% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_rawwater}
\alias{scrape_rawwater}
\title{Scrape raw water data}
\usage{
scrape_rawwater(spreadsheet, clearwell_start, labdat_parameters)
}
\arguments{
\item{spreadsheet}{dataframe. All data in the Weekly sheet of the lab data file}

\item{clearwell_start}{numeric value. The start of the Clearwell data as
indicated by the phrase "CLEAR WELL" in the Parameters column of the Weekly
data}

\item{labdat_parameters}{dataframe. Slightly processed parameters.xlsx data}
}
\value{
data frame of the raw water data
}
\description{
Read in raw water data from the Weekly (1st) sheet of the lab data file.
Identify parameters to read in based on parameters.xlsx
}
