% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_calculations.R
\name{summarize_monthly}
\alias{summarize_monthly}
\title{Summarize data by month}
\usage{
summarize_monthly(labdat)
}
\arguments{
\item{labdat}{labdat data file}
}
\value{

}
\description{
Summarize data by month
}
