% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{summarize_monthly_table}
\alias{summarize_monthly_table}
\title{Build monthly summary table}
\usage{
summarize_monthly_table(
  df = "",
  grouping = c("operations", "physical", "majorConstituents", "traceConstituents", "THM",
    "biological", "bacteriological"),
  sampling_station = "",
  year_select = NULL
)
}
\arguments{
\item{df}{dataframe. Data for table, Default: ''}

\item{grouping}{string. data grouping, Default:
c("operation", "physical", "majorConstituents", "traceConstituents",
"THM", "biological", "bacteriological")}

\item{sampling_station}{string. Sampling location, Default: ''}

\item{year_select}{numeric. Desired year, Default: NULL}
}
\value{
dataframe
}
\description{
Build monthly summary table
}
