% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{summarize_THM}
\alias{summarize_THM}
\title{TSummarize trihalomethanes for figure}
\usage{
summarize_THM(labdat)
}
\arguments{
\item{labdat}{}
}
\value{

}
\description{
TSummarize trihalomethanes for figure
}
