% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{suva}
\alias{suva}
\title{Calculate SUVA}
\usage{
suva(column_names, labdat)
}
\arguments{
\item{column_names}{string vector defined in apply_calculations. The desired final
columns for each calculated dataframe}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
numeric vector. The calculated values
}
\description{
Calculate SUVA, equal to (UV254)*100/DOC
}
