% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{suva_coag_filt}
\alias{suva_coag_filt}
\title{Calculate SUVA - Coagulation, Filtration}
\usage{
suva_coag_filt(column_names, labdat)
}
\arguments{
\item{column_names}{constant string vector. The desired column names in order}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate SUVA - Coagulation, Filtration, equal to:
\itemize{
\item (PreGAC UV254)\*100/PreGAC DOC if the PreGAC UV and DOC values exist
\item (Clearwell UV254)\*100/Clearwell DOC if the PreGAC UV and DOC values do not
exist
}

Blair decided that this value is no longer required. Code kept in case if
it is required in future
}
