% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_bpwtp}
\alias{theme_bpwtp}
\title{BPWTP theme}
\usage{
theme_bpwtp(base_size = 8, base_family = "")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\value{
theme
}
\description{
BPWTP theme
}
