% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{TOCremoval_coag}
\alias{TOCremoval_coag}
\title{Calculate TOC removal by coagulation}
\usage{
TOCremoval_coag(labdat)
}
\arguments{
\item{labdat}{routine labdat dataset}
}
\value{
calculated odour removal by coagulation
}
\description{
Calculate TOC removal by coagulation
}
\examples{
 #DOCremoval_GACfilt(labdat)
}
