% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{tot_B_BG_algae}
\alias{tot_B_BG_algae}
\title{Calculate total B+BG algae}
\usage{
tot_B_BG_algae(column_names, labdat)
}
\arguments{
\item{column_names}{string vector defined in apply_calculations. The desired final
columns for each calculated dataframe}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
numeric vector. The calculated values
}
\description{
Calculate the total blue plus blue green algae, equal blue green algae +
green algae
}
