% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{tot_chlorine_dose}
\alias{tot_chlorine_dose}
\title{Calculate total chlorine dose
Calculate the total chlorine dose, equal to the sum of chlorine pre,
intermed, and post}
\usage{
tot_chlorine_dose(column_names, labdat)
}
\arguments{
\item{column_names}{string vector defined in apply_calculations. The desired final
columns for each calculated dataframe}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate total chlorine dose
Calculate the total chlorine dose, equal to the sum of chlorine pre,
intermed, and post
}
