% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{tot_chlorineDose}
\alias{tot_chlorineDose}
\title{Total Chlorine Dose Calculation (Raw Water)}
\usage{
tot_chlorineDose(labdat)
}
\arguments{
\item{labdat}{Weekly routine sampling data.}
}
\value{
Dataframe containing datasheet (RawWater), station (Raw), datetimes,
parameter (tot.chlorineDose_mg.L), and result (Total Chlorine dose).
}
\description{
Calculates the sum of Chlorine-pre, Chlorine-intermed, and Chlorine-post.
}
\details{
Total Chlorine dose = Chlorine-pre + Chlorine-intermed + Chlorine-post.
}
\examples{
tot_chlorineDose(labdat = labdat)

}
