% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{total_THMs}
\alias{total_THMs}
\title{Calculate the total THMs
Calculate the total THMs, equal to the sum of bromodichloromethane,
bromoform, chlorodibromomethane, chloroform}
\usage{
total_THMs(column_names, labdat)
}
\arguments{
\item{column_names}{constant string vector. The desired column names in order}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate the total THMs
Calculate the total THMs, equal to the sum of bromodichloromethane,
bromoform, chlorodibromomethane, chloroform
}
