% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{trend_fig}
\alias{trend_fig}
\title{Plot parameter yearly trend}
\usage{
trend_fig(df = "", year_select = "", var = "")
}
\arguments{
\item{df}{data frame}

\item{year_select}{year of interest}

\item{var}{variable of interest}
}
\value{

}
\description{
Plot parameter yearly trend
}
