% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{turbidity_logRemoval}
\alias{turbidity_logRemoval}
\title{Calculate turbidity log removal}
\usage{
turbidity_logRemoval(labdat)
}
\arguments{
\item{labdat}{routine labdat file}
}
\value{
calculated turbidity log removal
}
\description{
Calculate turbidity log removal
}
