% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_database.R
\name{update_database}
\alias{update_database}
\title{Update database with data contained in the specified labdat file}
\usage{
update_database(path_to_labdat_file, path_to_db_file)
}
\arguments{
\item{path_to_labdat_file}{string. Path to the lab data file to be inserted
into the database file}

\item{path_to_db_file}{string. Path to the .csv database file if it exists,
and desired path to a .csv database file if it does not exist}
}
\value{
dataframe containing the processed lab data that has been inserted
into the database file
}
\description{
Update database with data contained in the specified labdat file
}
