% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_datafile}
\alias{write_datafile}
\title{Save a labdat data file}
\usage{
write_datafile(file, outdir, outfilename)
}
\arguments{
\item{outfilename}{}
}
\value{

}
\description{
Save a labdat data file
}
