% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{append_calc_values}
\alias{append_calc_values}
\title{Add calculated values to dataset}
\usage{
append_calc_values(labdat, labdat_calcs)
}
\arguments{
\item{labdat_calcs}{}
}
\value{

}
\description{
Add calculated values to dataset
}
