% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{apply_calculations}
\alias{apply_calculations}
\title{Apply weekly calculations}
\usage{
apply_calculations(labdat)
}
\arguments{
\item{labdat}{routine lab dat file}
}
\value{
calculated values
}
\description{
Apply weekly calculations
}
