% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{convert_adl}
\alias{convert_adl}
\title{Replace above detection limit values}
\usage{
convert_adl(labdat, replacement = NA)
}
\arguments{
\item{labdat}{labdat datafile}

\item{replacement}{value to replace adl values (default = NA)}
}
\value{

}
\description{
Replace above detection limit values
}
