% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrections.R
\name{convert_biocounts}
\alias{convert_biocounts}
\title{Convert biological data}
\usage{
convert_biocounts(labdat)
}
\arguments{
\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing the parsed lab data with the required biological
data values and units converted to per L as required
}
\description{
Identifies biological values stored in per mL by identifying \code{10^-3} in the
parameters name and converts the values to per L by multiplying by 1000.
}
