% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{DO_percent}
\alias{DO_percent}
\title{Bench Dissolved Oxygen (\%) (Percent Saturation DO) Calculation (Raw Water & Clearwell)}
\usage{
DO_percent(labdat, datadir = "data/data_tables", O2table_file = "O2Table.csv")
}
\arguments{
\item{labdat}{Weekly routine sampling data.}

\item{O2table_file}{Table of oxygen solubilities in water.}
}
\value{
Dataframe containing datasheet (RawWater, ClearWell), station (Raw,
Clearwell), datetimes, parameter (DO_percent), and result (DO
concentration expressed as a percentage).
}
\description{
Calculates dissolved oxygen concentration (\%).
Measured dissolved oxygen concentration (mg/L) is divided by the
saturation concentration of oxygen at the water temperature the sample
was collected at. This value is multiplied by 100 to express as a
percentage. The saturation of DO is determined from an oxygen solubility
table.
}
\details{
Dissolved Oxygen (\%) = Bench Diss. Oxygen (mg/L) ÷ Oxygen Solubility (mg/L) * 100.
}
\examples{
calc_BenchDO_percent(O2_table = O2_table, labdat = labdat)

}
