% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{group_table_historical}
\alias{group_table_historical}
\title{Build table with historical values}
\usage{
group_table_historical(
  df = "",
  year_select = "",
  grouping = "",
  sampling_station = "",
  percentiles = "",
  by_month = T,
  month_select = NULL
)
}
\arguments{
\item{month_select}{}
}
\value{

}
\description{
Build table with historical values
}
