% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{group_table_historical}
\alias{group_table_historical}
\title{Build table with historical values}
\usage{
group_table_historical(
  df = "",
  year_select = "",
  grouping = "",
  sampling_station = "",
  percentiles = "",
  by_month = T,
  month_select = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{year_select}{year of interest}

\item{grouping}{operational grouping}

\item{sampling_station}{sampling location}

\item{percentiles}{percentiles of interest}

\item{by_month}{group by month (T) or by year (F)}

\item{month_select}{month of interest}
}
\value{

}
\description{
Build table with historical values
}
