% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_calculations.R
\name{historical_basics}
\alias{historical_basics}
\title{Calculate historical stats}
\usage{
historical_basics(df = "", sampling_station = "", grouping = "", by_month = T)
}
\arguments{
\item{df}{dataframe for analysis}

\item{sampling_station}{sampling location}

\item{grouping}{operational metric}

\item{by_month}{collate data by month (T) or by year (F)}
}
\value{

}
\description{
Calculate historical stats
}
