% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_calculations.R
\name{historical_basics}
\alias{historical_basics}
\title{TCalculate historical stats}
\usage{
historical_basics(df = "", sampling_station = "", grouping = "", by_month = T)
}
\arguments{
\item{by_month}{}
}
\value{

}
\description{
TCalculate historical stats
}
