% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_calculations.R
\name{historical_extremes}
\alias{historical_extremes}
\title{Calculate historical extremes}
\usage{
historical_extremes(
  df = "",
  sampling_station = "",
  grouping = "",
  by_month = T
)
}
\arguments{
\item{df}{dataframe for analysis}

\item{sampling_station}{sampling location}

\item{grouping}{operational metric}

\item{by_month}{collate data by month (T) or by year (F)}
}
\value{
dataframe. Historical extremes
}
\description{
Calculate historical extremes
}
