% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_calculations.R
\name{historical_percentiles}
\alias{historical_percentiles}
\title{Calculate historical percentages}
\usage{
historical_percentiles(
  df = "",
  percentiles = c(0.1, 0.25, 0.5, 0.75, 0.9),
  sampling_station = "",
  grouping = "",
  by_month = T
)
}
\arguments{
\item{df}{dataframe for analysis}

\item{percentiles}{vector. Desired percentiles. Defaults to c(0.1, .25, .50, .75, .9)}

\item{sampling_station}{sampling location}

\item{grouping}{string. operational metric. parm_tag to filter by}

\item{by_month}{collate data by month (T) or by year (F)}
}
\value{
dataframe. Historical percentages
}
\description{
Calculate historical percentages
}
