% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{overall_coag_dose}
\alias{overall_coag_dose}
\title{Calculate the overall coagulant dose}
\usage{
overall_coag_dose(column_names, labdat, dry_alum_equiv_values)
}
\arguments{
\item{column_names}{constant string vector. The desired column names in order}

\item{labdat}{dataframe containing the parsed lab data}

\item{dry_alum_equiv_values}{numeric vector. The values calculated by
\code{PACl_DAE()}}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate the overall coagulant dose, equal to:
\itemize{
\item PACl DAE, if both train values exist
\item (PACl Train X\*2.07)+Alum)/2, if only one train value exists
\item Alum, if neither train value exists
}
}
