% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{PACl}
\alias{PACl}
\title{Calculate PACl}
\usage{
PACl(labdat)
}
\arguments{
\item{labdat}{dataframe containing the parsed lab data}
}
\value{
numeric vector. The calculated values
}
\description{
Calculate PACl, equal to (PACl Train A + PACl Train B)/2
}
