% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{percent_yield}
\alias{percent_yield}
\title{Calculate percent yield}
\usage{
percent_yield(pre, post)
}
\arguments{
\item{pre}{first measurement}

\item{post}{last measurement}
}
\value{

}
\description{
Calculate percent yield
}
\examples{
#percent_yield()
}
