% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{plot_all_parameters}
\alias{plot_all_parameters}
\title{Plot all possible parameters}
\usage{
plot_all_parameters(
  df,
  year_select = "",
  parameter_group = c("physical", "biological", "bacteriological",
    "major_constituents", "trace_constituents", "operations", "all"),
  station_list = c("Raw", "PreFM", "Channel", "PreGAC", "Clearwell", "all")
)
}
\arguments{
\item{df}{labdat dataset}

\item{year_select}{year of interest}

\item{parameter_group}{data grouping}

\item{station_list}{station}
}
\value{

}
\description{
Plot all possible parameters
}
\examples{
#
}
