% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_db}
\alias{read_db}
\title{Read the csv containing the database}
\usage{
read_db(path_to_db_file)
}
\arguments{
\item{path_to_db_file}{character string. Path to the file that contains the
database file, or if the file does not yet exist, the desired path to the
file that will contain the database file}
}
\value{
dataframe containing the database, if it exists, or NULL if it does not
}
\description{
Read the csv containing all of the processed labdat data, if it exists. If it
does not exist, it will later be created
}
