% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{removal_GACfilt}
\alias{removal_GACfilt}
\title{Calculate removal by GAC filtration}
\usage{
removal_GACfilt(column_names, labdat, parameter)
}
\arguments{
\item{column_names}{constant string vector. The desired column names in order}

\item{labdat}{dataframe containing the parsed lab data}

\item{parameter}{string. Parameter for which to calculate removal by
coagulation, filtration. Likely one of: DOC, TOC, Odour}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate removal by GAC filtration. Equal to percent yield between the
PreGAC and Clearwell values
}
