% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_clearwell}
\alias{scrape_clearwell}
\title{Scrape clearwell data from \code{.xls} file}
\usage{
scrape_clearwell(labdat_filename)
}
\arguments{
\item{labdat_filename}{file name for routine lab data \code{.xls} file}
}
\value{
scraped clearwell data
}
\description{
Scrape clearwell data from \code{.xls} file
}
\examples{
# scrape_clearwell(labdat_filename = "routinelabdata_2019.xls")
}
