% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_clearwell}
\alias{scrape_clearwell}
\title{Scrape Clearwell data}
\usage{
scrape_clearwell(spreadsheet, clearwell_start, labdat_parameters)
}
\arguments{
\item{spreadsheet}{dataframe. All data in the Weekly sheet of the lab data file}

\item{clearwell_start}{numeric value. The start of the Clearwell data as
indicated by the phrase "CLEAR WELL" in the Parameters column of the Weekly
data}

\item{labdat_parameters}{dataframe. Slightly processed parameters.xlsx data}
}
\value{
data frame of the Clearwell data (excluding THMs and Al)
}
\description{
Read in Clearwell data (excluding THM and Al data). Identify parameters to
read in based on parameters.xlsx
}
