% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_clearwell_al}
\alias{scrape_clearwell_al}
\title{Scrape clearwell aluminum data from \code{.xls} file}
\usage{
scrape_clearwell_al(labdat_filename)
}
\arguments{
\item{labdat_filename}{file name for routine lab data \code{.xls} file}
}
\value{
scraped clearwell aluminum data
}
\description{
Scrape clearwell aluminum data from \code{.xls} file
}
\examples{
# scrape_clearwell_al(labdat_filename = "routinelabdata_2019.xls")
}
