% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_clearwell_thms}
\alias{scrape_clearwell_thms}
\title{Scrape clearwell total halomethanes (thms) data from \code{.xls} file}
\usage{
scrape_clearwell_thms(labdat_filename, clearwell_range)
}
\arguments{
\item{labdat_filename}{file name for routine lab data \code{.xls} file}

\item{clearwell_range}{excel range for clearwell data}
}
\value{
scraped clearwell thms data
}
\description{
Scrape clearwell total halomethanes (thms) data from \code{.xls} file
}
\examples{
# scrape_clearwell_thms(labdat_filename = "routinelabdata_2019.xls")
}
