% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_docprofiles}
\alias{scrape_docprofiles}
\title{Scrape DOC profiles}
\usage{
scrape_docprofiles(path_to_labdat_file)
}
\arguments{
\item{path_to_labdat_file}{string. Path to the lab data file to be inserted
into the database file}
}
\value{
data frame of the DOC profile data
}
\description{
Read in DOC profile data
}
