% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_docprofiles}
\alias{scrape_docprofiles}
\title{Scrape DOC profiles}
\usage{
scrape_docprofiles(doc_data, labdat_parameters)
}
\arguments{
\item{doc_data}{dataframe. Output from read_doc()}

\item{labdat_parameters}{dataframe. Output from check_parameters_content}
}
\value{
data frame of the DOC profile data
}
\description{
Read in DOC profile data
}
