% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_labdatxls}
\alias{scrape_labdatxls}
\title{Scrape data from BPWTP routine lab data (\code{.xls})}
\usage{
scrape_labdatxls(labdat_filename, save_output = FALSE)
}
\arguments{
\item{labdat_filename}{file name for routine lab data \code{.xls} file}

\item{save_output}{Toggled option to save data output}
}
\value{
data frame of scraped data from an individual sheet
}
\description{
Scrape data from BPWTP routine lab data (\code{.xls})
}
\examples{
# scrape_labdatxls(labdat_filename = "routinlabdat_2019.xls", save_output = FALSE)
}
