% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_labdatxls}
\alias{scrape_labdatxls}
\title{Scrape data from BPWTP routine lab data}
\usage{
scrape_labdatxls(path_to_labdat_file, path_to_parameters)
}
\arguments{
\item{path_to_labdat_file}{string. Path to the lab data file to be inserted
into the database file}

\item{path_to_parameters}{string. Path to the .xlsx file that contains the
desired parameters and associated info (station, units, corrected units...)}
}
\value{
data frame of all the data contained in the Weekly lab data file
}
\description{
Read in all data contained in the Weekly (1st) sheet of the lab data file.
}
