% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape_rawwater}
\alias{scrape_rawwater}
\title{Scrape raw water data from \code{.xls} file}
\usage{
scrape_rawwater(labdat_filename, rawwater_range = "A8:BE130")
}
\arguments{
\item{labdat_filename}{file name for routine lab data \code{.xls} file}
}
\value{
scraped raw water data
}
\description{
Scrape raw water data from \code{.xls} file
}
\examples{
# scrape_rawwater(labdat_filename = "routinelabdata_2019.xls")
}
