% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{summarize_monthly_table}
\alias{summarize_monthly_table}
\title{Build monthly summary table}
\usage{
summarize_monthly_table(
  df = "",
  grouping = c("operation", "physical", "majorConstituents", "traceConstituents",
    "THM", "biological", "bacteriological"),
  sampling_station = "",
  year_select = NULL
)
}
\arguments{
\item{df}{data for table}

\item{grouping}{data grouping}

\item{sampling_station}{sampling location}

\item{year_select}{year of interest}
}
\value{

}
\description{
Build monthly summary table
}
\examples{
#
}
