% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_calculations.R
\name{summarize_THM}
\alias{summarize_THM}
\title{Summarize trihalomethanes for figure TBD}
\usage{
summarize_THM(labdat)
}
\arguments{
\item{labdat}{labdat data file}
}
\value{
dataframe. THM data summarized
}
\description{
Summarize trihalomethanes for figure TBD
}
