% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{tot_B_BG_algae}
\alias{tot_B_BG_algae}
\title{Sum total blue green algae and green algae}
\usage{
tot_B_BG_algae(labdat)
}
\arguments{
\item{labdat}{routine labdat}
}
\value{
summed blue greens and greens
}
\description{
Sum total blue green algae and green algae
}
