% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{tthm_fig}
\alias{tthm_fig}
\title{tTHM figure}
\usage{
tthm_fig(df = "thms", year_select = "")
}
\arguments{
\item{df}{default is thms}

\item{year_select}{year of interest}
}
\value{

}
\description{
tTHM figure
}
