% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{turbidity_logRemoval}
\alias{turbidity_logRemoval}
\title{Calculate the turbidity log removal
Calculate the turbidity log removal, equal to
log10(raw turbidity/Clearwell turbidity)}
\usage{
turbidity_logRemoval(column_names, labdat)
}
\arguments{
\item{column_names}{string vector defined in apply_calculations. The desired final
columns for each calculated dataframe}

\item{labdat}{dataframe containing the parsed lab data}
}
\value{
dataframe containing all required columns, including the calculated
result. Note that the results of \code{determine_NA_NC()} are employed, and that
parameters for which a value cannot be calculated are assigned NA or NaN, as
defined in the function description
}
\description{
Calculate the turbidity log removal
Calculate the turbidity log removal, equal to
log10(raw turbidity/Clearwell turbidity)
}
