% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_database.R
\name{update_database}
\alias{update_database}
\title{Update database with new data}
\usage{
update_database(
  labdat_file = "BPWTP_routinelabdata_historicalbase_clean.csv",
  datadir,
  xlsdir = "labdat_datafiles",
  outdir = "./data",
  outfilename = "BPWTP_labdat_updated.csv"
)
}
\arguments{
\item{outfilename}{}
}
\value{

}
\description{
Update database with new data
}
\examples{
#update_database()
}
