% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_database.R
\name{update_database}
\alias{update_database}
\title{Update database with data contained in the specified labdat file}
\usage{
update_database(
  path_to_labdat_file,
  path_to_db_file,
  path_to_db_dates_file,
  path_to_parameters,
  file_sheet_year = NULL
)
}
\arguments{
\item{path_to_labdat_file}{string. Full path (from root directory) to the lab
data file to be inserted into the database file}

\item{path_to_db_file}{string. Full path (from root directory) to the .csv
database file if it exists, and desired path to a .csv database file if it
does not exist}

\item{path_to_db_dates_file}{string. Full path (from root directory) to the .csv
database dates file if it exists, and desired path to a .csv database dates file if it
does not exist}

\item{path_to_parameters}{string. Full path (from root directory) to the
.xlsx file that contains the desired parameters and associated info
(station, units, corrected units...)}

\item{file_sheet_year}{numeric. The year of the labdata file of interest}
}
\value{
dataframe containing the processed lab data that has been newly
inserted into the database file
}
\description{
Update database with data contained in the specified labdat file
}
